requirements.txt
numpy==1.26.2
pygame==2.5.2
py-sudoku==1.0.3